Refactor FilterCommand predicate
The FilterCommand predicate was of
type NameContainsKeyphrasePredicate.

This modification allows the command
to apply a wider variety of conditions
for filtering Person objects,
thereby increasing its utility in various contexts.

Refactor the FilterCommand within the application
to utilize a generic Predicate<Person> interface.

The change facilitates the introduction of
more complex filtering criteria
without necessitating additional modifications
to the FilterCommand's underlying implementation.